Rework Md API
The changes presented here make Md a non-trusting and safe API by being
private and following a Date-like getter and setter system. The changes
in the commit solidify the naming and style of usage and also prevent
false values so Md is no longer a simple tuple and is more akin to a
token struct, a helper for intermediate value conversions.